Release 6.4.0

+++ module_core/src/main/java/com/sensorsdata/analytics/android/sdk/autotrack/ActivityPageLeaveCallbacks.java
@@ -161,7 +161,7 @@ public class ActivityPageLeaveCallbacks implements SensorsDataActivityLifecycleC
         try {
             long resumeTime = properties.optLong(START_TIME);
             properties.remove(START_TIME);
-            double duration = TimeUtils.duration(resumeTime, SystemClock.elapsedRealtime());
+            Float duration = TimeUtils.duration(resumeTime, SystemClock.elapsedRealtime());
             if (duration < 0.05) {
                 return;
             }
